now that we have fewer globals, do most initialization in xrun's constructor
