+++ include/threadinfo.hh
@@ -52,6 +52,23 @@ public:
   }
 
   void initialize() {
+      struct rlimit rl;
+
+      // Get the stack size.
+      if(Real::getrlimit(RLIMIT_STACK, &rl) != 0) {
+        PRWRN("Get the stack size failed.\n");
+        Real::exit(-1);
+      }
+
+      // if there is no limit for our stack size, then just pick a
+      // reasonable limit.
+      if (rl.rlim_cur == (rlim_t)-1) {
+        rl.rlim_cur = 2048*4096; // 8 MB
+      }
+
+      __max_stack_size = rl.rlim_cur;
+
+
     _aliveThreads = 0;
     _reapableThreads = 0;
     _threadIndex = 0;
